Refactor MainForm into partial classes
